+++ Pandora/Paradigm/Structure/Some/Splay.hs
@@ -6,14 +6,14 @@ import Pandora.Pattern.Category ((.), ($))
 import Pandora.Pattern.Functor.Covariant (Covariant ((<$>)))
 import Pandora.Pattern.Functor.Extractable (extract)
 import Pandora.Pattern.Functor.Bindable (Bindable ((>>=)))
-import Pandora.Paradigm.Primary.Functor (left, right, branches)
+import Pandora.Paradigm.Primary.Functor (branches)
 import Pandora.Paradigm.Primary.Functor.Function ((%))
 import Pandora.Paradigm.Primary.Functor.Maybe (Maybe (Just))
 import Pandora.Paradigm.Primary.Functor.Wye (Wye (Left, Right))
 import Pandora.Paradigm.Primary.Transformer.Construction (Construction (Construct), deconstruct)
 import Pandora.Paradigm.Controlflow.Effect.Interpreted (run, (||=))
 import Pandora.Paradigm.Schemes (TU (TU))
-import Pandora.Paradigm.Structure.Ability.Morphable (Morphable (Morphing, morphing), Morph (Rotate), rotate)
+import Pandora.Paradigm.Structure.Ability.Morphable (Morphable (Morphing, morphing), Morph (Rotate), rotate, into)
 import Pandora.Paradigm.Structure.Ability.Substructure (substitute)
 import Pandora.Paradigm.Structure.Some.Binary (Binary)
 
@@ -21,17 +21,18 @@ data Splay a = Zig a | Zag a
 
 instance Morphable (Rotate (Left Zig)) (Construction Wye) where
 	type Morphing (Rotate (Left Zig)) (Construction Wye) = Binary
-	morphing (extract . run -> Construct parent st) = TU $ Construct % subtree . extract <$> left st where
+	morphing (extract . run -> Construct parent st) = TU $ Construct % subtree . extract <$> into @(Left Maybe) st where
 
-		subtree = branches (deconstruct <$> left st >>= left) . Just . Construct parent
-			$ branches (deconstruct <$> left st >>= right) (right st)
+		subtree = branches (deconstruct <$> into @(Left Maybe) st >>= into @(Left Maybe)) . Just . Construct parent
+			$ branches (deconstruct <$> into @(Left Maybe) st >>= into @(Right Maybe)) (into @(Right Maybe) st)
 
 instance Morphable (Rotate (Right Zig)) (Construction Wye) where
 	type Morphing (Rotate (Right Zig)) (Construction Wye) = Binary
-	morphing (extract . run -> Construct parent st) = TU $ Construct % subtree . extract <$> right st where
+	morphing (extract . run -> Construct parent st) = TU $ Construct % subtree . extract <$> into @(Right Maybe) st where
 
-		subtree = branches (left st) . Just . Construct parent
-			$ branches (deconstruct <$> right st >>= left) (deconstruct <$> right st >>= right)
+		subtree = branches (into @(Left Maybe) st) . Just . Construct parent
+			$ branches (deconstruct <$> into @(Right Maybe) st >>= into @(Left Maybe))
+				(deconstruct <$> into @(Right Maybe) st >>= into @(Right Maybe))
 
 instance Morphable (Rotate (Left (Zig Zig))) (Construction Wye) where
 	type Morphing (Rotate (Left (Zig Zig))) (Construction Wye) = Binary